Extend grammar
It now accepts both finite automata and turing machines. I don't
know what is the best way to organize the parser, we have the
option to agroup all functions related to some type of machine.
There are some warnings but I still didn't figure it out how to remove them.